+++ src/modules/authentication/screens/index.ts
@@ -1,0 +1,3 @@
+export { SplashScreen } from './SplashScreen/SplashScreen';
+export { LoginScreen } from './LoginScreen/LoginScreen';
+export { SignUpScreen } from './SignUpScreen/SignUpScreen';